api: allow to configure the API endpoint, without running the API service
Some handlers (like the HipChat handler) require access to the API server
to discover if the event is part of a stash or not
(see https://github.com/sensu-plugins/sensu-plugin/blob/master/lib/sensu-handler.rb#L96)
for more information.

This commit splits the configuration part from the service part, which
allows the configuration file to be reused across states.